requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
distlib==0.3.6
Django==4.1.2
filelock==3.8.0
idna==3.4
numpy==1.23.4
Pillow==9.2.0
platformdirs==2.5.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tflearn==0.5.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
tqdm==4.64.0
typing_extensions==4.2.0
tzdata==2021.5
urllib3==1.24.2
virtualenv==20.13.0
virtualenvwrapper-win==1.2.7
waitress==2.0.0
Werkzeug==2.1.2
whitenoise==3.3.1
wrapt==1.14.1
ypstruct==0.0.2
